edit ImageAlbum

+++ topalbum/TopAlbumsAdapter.kt
@@ -1,20 +1,26 @@
 package ir.android.musicmanager.topalbum
 
+import android.app.Activity
+import android.app.AlertDialog
 import android.content.Context
+import android.content.DialogInterface
 import android.content.Intent
+import android.content.pm.PackageManager
+import android.net.Uri
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
-import ir.android.musicmanager.pojos.topAlbum.Album
-import ir.android.musicmanager.pojos.topAlbum.AlbumDetailDataBase
-import kotlinx.android.synthetic.main.album_item.view.*
-
-import android.net.Uri
 import com.squareup.picasso.Picasso
 import ir.android.musicmanager.R
+import ir.android.musicmanager.pojos.topAlbum.Album
+import ir.android.musicmanager.pojos.topAlbum.AlbumDetailDataBase
 import ir.android.musicmanager.utils.SaveImage
+import kotlinx.android.synthetic.main.album_item.view.*
 
 
 @Suppress("DEPRECATION")
@@ -22,7 +28,8 @@ import ir.android.musicmanager.utils.SaveImage
  * Created by  Roxa on 5/11/2019.
  */
 class TopAlbumsAdapter(val topAlbumList: List<Album>, clickHandler: ClickHandler) :
-    RecyclerView.Adapter<TopAlbumsAdapter.TopAlbumsViewHolder>() {
+    RecyclerView.Adapter<TopAlbumsAdapter.TopAlbumsViewHolder>(), ActivityCompat.OnRequestPermissionsResultCallback {
+
     val clickHandler = clickHandler
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TopAlbumsViewHolder {
 
@@ -41,6 +48,7 @@ class TopAlbumsAdapter(val topAlbumList: List<Album>, clickHandler: ClickHandler
     class TopAlbumsViewHolder(val view: View, val clickHandler: ClickHandler, val context: Context) :
         RecyclerView.ViewHolder(view) {
 
+
         fun bindItems(imdb: Album) {
 
             view.txt_album_artist.text = imdb.artist.name
@@ -48,6 +56,7 @@ class TopAlbumsAdapter(val topAlbumList: List<Album>, clickHandler: ClickHandler
 
             if (imdb.image[2].text.isNotEmpty()) {
                 Picasso.with(context).load(imdb.image[2].text).into(view.img_album)
+                // Picasso.with(context).load(imdb.image[2].text).into(SaveImage.getTarget(imdb.mbid))
 
             }
 
@@ -61,16 +70,54 @@ class TopAlbumsAdapter(val topAlbumList: List<Album>, clickHandler: ClickHandler
             }
 
             view.btn_add.setOnClickListener {
+                checkPermission()
                 val dbHandler = DataBaseHelper(context, null)
-                val albums = AlbumDetailDataBase(imdb.artist.name, imdb.name, imdb.image[2].text)
+                val albums = AlbumDetailDataBase(imdb.artist.name, imdb.name, imdb.artist.mbid)
+
+                Picasso.with(context).load(imdb.image[2].text).into(SaveImage.getTarget(imdb.artist.mbid))
                 dbHandler.addAlbum(albums)
                 Toast.makeText(context, " Add to Archive ", Toast.LENGTH_SHORT).show()
             }
             view.btn_delete.setOnClickListener {
+                checkPermission()
                 val dbHandler = DataBaseHelper(context, null)
                 dbHandler.deleteAlbums()
                 Toast.makeText(context, " Delete from Archive ", Toast.LENGTH_SHORT).show()
             }
         }
+
+        fun checkPermission() {
+
+            val permission =
+                ContextCompat.checkSelfPermission(context, android.Manifest.permission.WRITE_EXTERNAL_STORAGE)
+            if (permission != PackageManager.PERMISSION_GRANTED) {
+                makeRequest()
+            }
+        }
+
+        fun makeRequest() {
+            ActivityCompat.requestPermissions(
+                context as Activity,
+                arrayOf(android.Manifest.permission.WRITE_EXTERNAL_STORAGE),
+                1
+            )
+        }
+    }
+
+
+    override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<out String>, grantResults: IntArray) {
+        if (grantResults.isEmpty() || grantResults[0] != PackageManager.PERMISSION_GRANTED) {
+
+            Log.d("permi", "errrrr")
+        }
+//        val listener = DialogInterface.OnClickListener { dialog, id -> finish() }
+//
+//        val builder = AlertDialog.Builder()
+//        builder.setTitle("Digikala")
+//            .setMessage("")
+//            .setPositiveButton("" ,listener)
+//            .show()
+
+
     }
 }
